[APIS-804] Add files in 10.2 and remove redundant files
